Tried to implement search in driver... still testing

+++ src/cmc/mario/controllers/SearchController.java
@@ -87,10 +87,10 @@ public class SearchController {
 	 *            an list of popular major or emphases at the university
 	 */
 	public List<University> search(String schoolName, String state, String location, String control,
-			Integer numOfStuStart, Integer numOfStuEnd, Double perFemStart, Double perFemEnd, Integer satVerbalStart,
+			Integer numOfStuStart, Integer numOfStuEnd, Integer perFemStart, Integer perFemEnd, Integer satVerbalStart,
 			Integer satVerbalEnd, Integer satMathStart, Integer satMathEnd, Integer priceStart, Integer priceEnd,
-			Integer finAidStart, Integer finAidEnd, Integer numOfAppStart, Integer numOfAppEnd, Double perAdmitStart,
-			Double perAdmitEnd, Double perEnrollStart, Double perEnrollEnd, Integer academicScaleStart,
+			Integer finAidStart, Integer finAidEnd, Integer numOfAppStart, Integer numOfAppEnd, Integer perAdmitStart,
+			Integer perAdmitEnd, Integer perEnrollStart, Integer perEnrollEnd, Integer academicScaleStart,
 			Integer academicScaleEnd, Integer socialScaleStart, Integer socialScaleEnd, Integer lifeScaleStart,
 			Integer lifeScaleEnd, List<String> popMajor) {
 		return dbControl.searchResults(schoolName, state, location, control, numOfStuStart, numOfStuEnd, perFemStart,
